changed int to float in player hp

+++ Assets/Scripts/Gameplay/ShipInfo.cs
@@ -1,9 +1,9 @@
 ﻿namespace STP.Gameplay {
     public readonly struct ShipInfo {
-        public readonly int   Hp;
+        public readonly float Hp;
         public readonly float MaxSpeed;
 
-        public ShipInfo(int hp, float maxSpeed) {
+        public ShipInfo(float hp, float maxSpeed) {
             Hp       = hp;
             MaxSpeed = maxSpeed;
         }
